+++ src/lib/arrow-redo.svelte
@@ -1,0 +1,33 @@
+<script lang="ts">
+	export let size = '24';
+	export let variation: 'regular' | 'outline' | 'sharp' = 'regular';
+	let svgpath: string;
+	let svgregular =
+		'<path d="M58.79,439.13A16,16,0,0,1,48,424c0-73.1,14.68-131.56,43.65-173.77,35-51,90.21-78.46,164.35-81.87V88a16,16,0,0,1,27.05-11.57l176,168a16,16,0,0,1,0,23.14l-176,168A16,16,0,0,1,256,424V344.23c-45,1.36-79,8.65-106.07,22.64-29.25,15.12-50.46,37.71-73.32,67a16,16,0,0,1-17.82,5.28Z"/>';
+	let svgoutline =
+		'<path d="M448,256,272,88v96C103.57,184,64,304.77,64,424c48.61-62.24,91.6-96,208-96v96Z" style="fill:none;stroke:#000;stroke-linejoin:round;stroke-width:32px"/>';
+	let svgsharp =
+		'<path d="M48,399.26C48,335.19,62.44,284,90.91,247c34.38-44.67,88.68-68.77,161.56-71.75V72L464,252,252.47,432V329.35c-44.25,1.19-77.66,7.58-104.27,19.84-28.75,13.25-49.6,33.05-72.08,58.7L48,440Z"/>';
+	switch (variation) {
+		case 'outline':
+			svgpath = svgoutline;
+			break;
+		case 'sharp':
+			svgpath = svgsharp;
+			break;
+		default:
+			svgpath = svgregular;
+	}
+</script>
+
+<svg
+	xmlns="http://www.w3.org/2000/svg"
+	width={size}
+	height={size}
+	class={$$props.class}
+	{...$$restProps}
+	aria-label={ariaLabel}
+	viewBox="0 0 512 512"
+>
+	{@html svgpath}
+</svg>