+++ src/lib/bar-chart.svelte
@@ -1,0 +1,33 @@
+<script lang="ts">
+	export let size = '24';
+	export let variation: 'regular' | 'outline' | 'sharp' = 'regular';
+	let svgpath: string;
+	let svgregular =
+		'<path d="M480,496H48a32,32,0,0,1-32-32V32a16,16,0,0,1,32,0V464H480a16,16,0,0,1,0,32Z"/><path d="M156,432H116a36,36,0,0,1-36-36V244a36,36,0,0,1,36-36h40a36,36,0,0,1,36,36V396A36,36,0,0,1,156,432Z"/><path d="M300,432H260a36,36,0,0,1-36-36V196a36,36,0,0,1,36-36h40a36,36,0,0,1,36,36V396A36,36,0,0,1,300,432Z"/><path d="M443.64,432h-40a36,36,0,0,1-36-36V132a36,36,0,0,1,36-36h40a36,36,0,0,1,36,36V396A36,36,0,0,1,443.64,432Z"/>';
+	let svgoutline =
+		'<path d="M32,32V464a16,16,0,0,0,16,16H480" style="fill:none;stroke:#000;stroke-linecap:round;stroke-linejoin:round;stroke-width:32px"/><rect x="96" y="224" width="80" height="192" rx="20" ry="20" style="fill:none;stroke:#000;stroke-linecap:round;stroke-linejoin:round;stroke-width:32px"/><rect x="240" y="176" width="80" height="240" rx="20" ry="20" style="fill:none;stroke:#000;stroke-linecap:round;stroke-linejoin:round;stroke-width:32px"/><rect x="383.64" y="112" width="80" height="304" rx="20" ry="20" style="fill:none;stroke:#000;stroke-linecap:round;stroke-linejoin:round;stroke-width:32px"/>';
+	let svgsharp =
+		'<polygon points="496 496 16 496 16 16 48 16 48 464 496 464 496 496"/><path d="M192,432H80V208H192Z"/><path d="M336,432H224V160H336Z"/><path d="M479.64,432h-112V96h112Z"/>';
+	switch (variation) {
+		case 'outline':
+			svgpath = svgoutline;
+			break;
+		case 'sharp':
+			svgpath = svgsharp;
+			break;
+		default:
+			svgpath = svgregular;
+	}
+</script>
+
+<svg
+	xmlns="http://www.w3.org/2000/svg"
+	width={size}
+	height={size}
+	class={$$props.class}
+	{...$$restProps}
+	aria-label={ariaLabel}
+	viewBox="0 0 512 512"
+>
+	{@html svgpath}
+</svg>