+++ src/lib/bowling-ball.svelte
@@ -1,0 +1,33 @@
+<script lang="ts">
+	export let size = '24';
+	export let variation: 'regular' | 'outline' | 'sharp' = 'regular';
+	let svgpath: string;
+	let svgregular =
+		'<path d="M414.39,97.61A224,224,0,1,0,97.61,414.39,224,224,0,1,0,414.39,97.61ZM288,224a24,24,0,1,1,24-24A24,24,0,0,1,288,224Zm8-72a24,24,0,1,1,24-24A24,24,0,0,1,296,152Zm64,40a24,24,0,1,1,24-24A24,24,0,0,1,360,192Z"/>';
+	let svgoutline =
+		'<circle cx="256" cy="256" r="208" fill="none" stroke="#000" stroke-miterlimit="10" stroke-width="32"/><circle cx="288" cy="200" r="24"/><circle cx="296" cy="128" r="24"/><circle cx="360" cy="168" r="24"/>';
+	let svgsharp =
+		'<path d="M414.39,97.61A224,224,0,1,0,97.61,414.39,224,224,0,1,0,414.39,97.61ZM286,230a28,28,0,1,1,28-28A28,28,0,0,1,286,230Zm8-76a28,28,0,1,1,28-28A28,28,0,0,1,294,154Zm68,44a28,28,0,1,1,28-28A28,28,0,0,1,362,198Z"/>';
+	switch (variation) {
+		case 'outline':
+			svgpath = svgoutline;
+			break;
+		case 'sharp':
+			svgpath = svgsharp;
+			break;
+		default:
+			svgpath = svgregular;
+	}
+</script>
+
+<svg
+	xmlns="http://www.w3.org/2000/svg"
+	width={size}
+	height={size}
+	class={$$props.class}
+	{...$$restProps}
+	aria-label={ariaLabel}
+	viewBox="0 0 512 512"
+>
+	{@html svgpath}
+</svg>