+++ src/lib/chevron-down-circle.svelte
@@ -1,0 +1,33 @@
+<script lang="ts">
+	export let size = '24';
+	export let variation: 'regular' | 'outline' | 'sharp' = 'regular';
+	let svgpath: string;
+	let svgregular =
+		'<path d="M464,256c0-114.87-93.13-208-208-208S48,141.13,48,256s93.13,208,208,208S464,370.87,464,256ZM363.31,227.31l-96,96a16,16,0,0,1-22.62,0l-96-96a16,16,0,0,1,22.62-22.62L256,289.37l84.69-84.68a16,16,0,0,1,22.62,22.62Z"/>';
+	let svgoutline =
+		'<path d="M256,64C150,64,64,150,64,256s86,192,192,192,192-86,192-192S362,64,256,64Z" style="fill:none;stroke:#000;stroke-miterlimit:10;stroke-width:32px"/><polyline points="352 216 256 312 160 216" style="fill:none;stroke:#000;stroke-linecap:round;stroke-linejoin:round;stroke-width:32px"/>';
+	let svgsharp =
+		'<path d="M256,464c114.87,0,208-93.13,208-208S370.87,48,256,48,48,141.13,48,256,141.13,464,256,464ZM160,193.37l96,96,96-96L374.63,216,256,334.63,137.37,216Z"/>';
+	switch (variation) {
+		case 'outline':
+			svgpath = svgoutline;
+			break;
+		case 'sharp':
+			svgpath = svgsharp;
+			break;
+		default:
+			svgpath = svgregular;
+	}
+</script>
+
+<svg
+	xmlns="http://www.w3.org/2000/svg"
+	width={size}
+	height={size}
+	class={$$props.class}
+	{...$$restProps}
+	aria-label={ariaLabel}
+	viewBox="0 0 512 512"
+>
+	{@html svgpath}
+</svg>